requirements.txt
earthengine-api
google-api-python-client
google-auth-httplib2
pandas
numpy
scikit-learn
tensorflow
fastapi
uvicorn[standard]
streamlit
requests
joblib

geopy